// resources/[soz]/soz-core/src/shared/draw.ts
export type Draw2dTextParameters = {
    x: number;
    y: number;
    width: number;
    height: number;
    scale: number;
    r: number;
    g: number;
    b: number;
    a: number;
    text: string;
};

export type Draw3dTextParameters = {
    x: number;
    y: number;
    z: number;
    text: string;
};